+++ html/iframePage/CHANGELOG.md
@@ -1,5 +1,40 @@
 # Build Changelog
 
+## v0.0.15
+
+- Built At: 2026-04-23T03:31:32.415Z
+- Previous Version: v0.0.14
+- Git Branch: master
+- Git Commit: c311f6b0
+
+### Changed Files
+- html/iframePage/CHANGELOG.md
+- html/iframePage/main/App.css
+- html/iframePage/main/App.tsx
+- html/iframePage/main/components/GroupWorkbench/index.tsx
+- html/iframePage/package-lock.json
+- html/iframePage/package.json
+- html/iframePage/stats.html
+- manifest.json
+
+
+## v0.0.14
+
+- Built At: 2026-04-22T16:00:47.708Z
+- Previous Version: v0.0.13
+- Git Branch: master
+- Git Commit: c311f6b0
+
+### Changed Files
+- html/iframePage/main/App.css
+- html/iframePage/main/App.tsx
+- html/iframePage/main/components/GroupWorkbench/index.tsx
+- html/iframePage/package-lock.json
+- html/iframePage/package.json
+- html/iframePage/stats.html
+- manifest.json
+
+
 ## v0.0.13
 
 - Built At: 2026-04-21T09:00:17.167Z